backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
flask-session==0.4.0
google-auth==2.16.2
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.80.0
python-dotenv==1.0.0
youtube-transcript-api==0.6.0
gunicorn==20.1.0
requests==2.28.2
werkzeug==2.2.3
celery
redis
elasticsearch==9.2.0